ci: Remove push event trigger for renovate
Renovate used to run both on a Cron schedule (hourly), as well as for every push. this caused some conflicts where one of the two would cancel the other (since they have a concurrency config to cancel an in progress run when a new run is triggered.

this situation sent me email notifications and it's annoying 

this commit removes the push event trigger and leaves only the Cron and manual triggers